refactor: move context to separated file
