+++ components/authentication/RegisterForm.js
@@ -2,7 +2,7 @@ import { Formik, Form, Field } from "formik";
 import axios from "axios";
 import PhoneInput from "./PhoneInput";
 import { useState } from "react";
-import Router from "next/router";
+import { useRouter } from "next/router";
 
 function validateEmail(value) {
   let error;
