+++ components/buttons/OrangeButtons.js
@@ -2,6 +2,7 @@ export function ButtonOrangeLight(props) {
   return (
     <button
       id={props.id}
+      type={props.type}
       className={`flex justify-center items-center py-3 px-6 bg-ps-orange-100 text-ps-orange-500 rounded-[99px] ${props.width} min-w-[120px] text-base font-bold text-nowrap`}
       onClick={props.onClick}
     >
@@ -14,6 +15,7 @@ export function ButtonOrange(props) {
   return (
     <button
       id={props.id}
+      type={props.type}
       className={`flex justify-center items-center py-3 px-6 bg-ps-orange-500 text-ps-white rounded-[99px] ${props.width} min-w-[120px]  text-base font-bold text-nowrap`} 
       onClick={props.onClick}
     >
