+++ components/layout.js
@@ -8,6 +8,7 @@ import Head from "next/head";
 import { SearchProvider } from "@/pages/context/Search";
 import { BookingProvider } from "@/pages/context/Booking";
 import { OnwerProvider } from "@/pages/context/Owners";
+import { SittersProvider } from "@/pages/context/SittersList";
 
 export default function Layout({ children }) {
   const [openModal, setOpenModal] = useState(false);
@@ -39,20 +40,22 @@ export default function Layout({ children }) {
       <OnwerProvider>
         <BookingProvider>
           <SearchProvider>
-            <div className="w-full">
-              {!noLayoutRoutes.includes(router.pathname) && (
-                <NavBar setOpenModal={() => setOpenModal((prev) => !prev)} />
-              )}
-              {openModal && (
-                <div className="absolute top-15 right-0 size-10 bg-ps-white w-full h-full z-10">
-                  <LoginMobile
-                    setOpenModal={() => setOpenModal((prev) => !prev)}
-                  />
-                </div>
-              )}
-              <div>{children}</div>
-              {!noLayoutRoutes.includes(router.pathname) && <Footer />}
-            </div>
+            <SittersProvider>
+              <div className="w-full">
+                {!noLayoutRoutes.includes(router.pathname) && (
+                  <NavBar setOpenModal={() => setOpenModal((prev) => !prev)} />
+                )}
+                {openModal && (
+                  <div className="absolute top-15 right-0 size-10 bg-ps-white w-full h-full z-10">
+                    <LoginMobile
+                      setOpenModal={() => setOpenModal((prev) => !prev)}
+                    />
+                  </div>
+                )}
+                <div>{children}</div>
+                {!noLayoutRoutes.includes(router.pathname) && <Footer />}
+              </div>
+            </SittersProvider>
           </SearchProvider>
         </BookingProvider>
       </OnwerProvider>
