+++ components/sitters/searchbar/CheckBoxSitter.js
@@ -1,9 +1,8 @@
-import { useContext } from "react";
-import { sittersContext } from "@/pages/sitters";
+import { useSitters } from "@/pages/context/SittersList";
 
 export default function CheckBoxSitter() {
   const { selectedPets, setSelectedPets, setPetQuery } =
-    useContext(sittersContext);
+    useSitters()
 
   function handlePetSelect(event) {
     const { value, checked } = event.target;
